Your game carries a lot of state that's very difficult to keep track of. Not sure why you send state related to the tile names and values around, seems like it would be a constant (it's also possible that I just don't understand your high level design).
For the unit tests I think the getInitalBoard should be sending a random set of tiles for each player but doesn't appear to be. Secondly I couldn't get an end of game scenario for placing tiles to pass. Not sure if this is a bug or there is an explicit call to end turn required. Again I had trouble testing these scenarios becuase of all the state but it seems that by keeping tile states in the state object that you could allow somebody to change the value of tile43 for example to a joker and you wouldn't know.